+++ Starward.Core/Hyperion/StarRail/Ledger/LedgerSummary.cs
@@ -10,7 +10,7 @@ public class LedgerSummary : IJsonOnDeserialized
 
     [JsonPropertyName("uid")]
     [JsonNumberHandling(JsonNumberHandling.AllowReadingFromString | JsonNumberHandling.WriteAsString)]
-    public int Uid { get; set; }
+    public long Uid { get; set; }
 
 
     [JsonPropertyName("region")]
